+++ sdk/network/arm-network/src/operations/networkInterfaceTapConfigurations.ts
@@ -14,12 +14,6 @@ import * as Mappers from "../models/mappers";
 import * as Parameters from "../models/parameters";
 import { NetworkManagementClient } from "../networkManagementClient";
 import {
-  SimplePollerLike,
-  OperationState,
-  createHttpPoller
-} from "@azure/core-lro";
-import { createLroSpec } from "../lroImpl";
-import {
   NetworkInterfaceTapConfiguration,
   NetworkInterfaceTapConfigurationsListNextOptionalParams,
   NetworkInterfaceTapConfigurationsListOptionalParams,
@@ -137,90 +131,21 @@ export class NetworkInterfaceTapConfigurationsImpl
    * @param tapConfigurationName The name of the tap configuration.
    * @param options The options parameters.
    */
-  async beginDelete(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    tapConfigurationName: string,
-    options?: NetworkInterfaceTapConfigurationsDeleteOptionalParams
-  ): Promise<SimplePollerLike<OperationState<void>, void>> {
-    const directSendOperation = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ): Promise<void> => {
-      return this.client.sendOperationRequest(args, spec);
-    };
-    const sendOperationFn = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
-      const providedCallback = args.options?.onResponse;
-      const callback: coreClient.RawResponseCallback = (
-        rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
-      ) => {
-        currentRawResponse = rawResponse;
-        providedCallback?.(rawResponse, flatResponse);
-      };
-      const updatedArgs = {
-        ...args,
-        options: {
-          ...args.options,
-          onResponse: callback
-        }
-      };
-      const flatResponse = await directSendOperation(updatedArgs, spec);
-      return {
-        flatResponse,
-        rawResponse: {
-          statusCode: currentRawResponse!.status,
-          body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
-      };
-    };
-
-    const lro = createLroSpec({
-      sendOperationFn,
-      args: {
-        resourceGroupName,
-        networkInterfaceName,
-        tapConfigurationName,
-        options
-      },
-      spec: deleteOperationSpec
-    });
-    const poller = await createHttpPoller<void, OperationState<void>>(lro, {
-      restoreFrom: options?.resumeFrom,
-      intervalInMs: options?.updateIntervalInMs,
-      resourceLocationConfig: "location"
-    });
-    await poller.poll();
-    return poller;
-  }
-
-  /**
-   * Deletes the specified tap configuration from the NetworkInterface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param tapConfigurationName The name of the tap configuration.
-   * @param options The options parameters.
-   */
-  async beginDeleteAndWait(
+  delete(
     resourceGroupName: string,
     networkInterfaceName: string,
     tapConfigurationName: string,
     options?: NetworkInterfaceTapConfigurationsDeleteOptionalParams
   ): Promise<void> {
-    const poller = await this.beginDelete(
-      resourceGroupName,
-      networkInterfaceName,
-      tapConfigurationName,
-      options
+    return this.client.sendOperationRequest(
+      {
+        resourceGroupName,
+        networkInterfaceName,
+        tapConfigurationName,
+        options
+      },
+      deleteOperationSpec
     );
-    return poller.pollUntilDone();
   }
 
   /**
@@ -256,104 +181,23 @@ export class NetworkInterfaceTapConfigurationsImpl
    *                                   operation.
    * @param options The options parameters.
    */
-  async beginCreateOrUpdate(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    tapConfigurationName: string,
-    tapConfigurationParameters: NetworkInterfaceTapConfiguration,
-    options?: NetworkInterfaceTapConfigurationsCreateOrUpdateOptionalParams
-  ): Promise<
-    SimplePollerLike<
-      OperationState<NetworkInterfaceTapConfigurationsCreateOrUpdateResponse>,
-      NetworkInterfaceTapConfigurationsCreateOrUpdateResponse
-    >
-  > {
-    const directSendOperation = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ): Promise<NetworkInterfaceTapConfigurationsCreateOrUpdateResponse> => {
-      return this.client.sendOperationRequest(args, spec);
-    };
-    const sendOperationFn = async (
-      args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
-    ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
-      const providedCallback = args.options?.onResponse;
-      const callback: coreClient.RawResponseCallback = (
-        rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
-      ) => {
-        currentRawResponse = rawResponse;
-        providedCallback?.(rawResponse, flatResponse);
-      };
-      const updatedArgs = {
-        ...args,
-        options: {
-          ...args.options,
-          onResponse: callback
-        }
-      };
-      const flatResponse = await directSendOperation(updatedArgs, spec);
-      return {
-        flatResponse,
-        rawResponse: {
-          statusCode: currentRawResponse!.status,
-          body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
-      };
-    };
-
-    const lro = createLroSpec({
-      sendOperationFn,
-      args: {
-        resourceGroupName,
-        networkInterfaceName,
-        tapConfigurationName,
-        tapConfigurationParameters,
-        options
-      },
-      spec: createOrUpdateOperationSpec
-    });
-    const poller = await createHttpPoller<
-      NetworkInterfaceTapConfigurationsCreateOrUpdateResponse,
-      OperationState<NetworkInterfaceTapConfigurationsCreateOrUpdateResponse>
-    >(lro, {
-      restoreFrom: options?.resumeFrom,
-      intervalInMs: options?.updateIntervalInMs,
-      resourceLocationConfig: "azure-async-operation"
-    });
-    await poller.poll();
-    return poller;
-  }
-
-  /**
-   * Creates or updates a Tap configuration in the specified NetworkInterface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param tapConfigurationName The name of the tap configuration.
-   * @param tapConfigurationParameters Parameters supplied to the create or update tap configuration
-   *                                   operation.
-   * @param options The options parameters.
-   */
-  async beginCreateOrUpdateAndWait(
+  createOrUpdate(
     resourceGroupName: string,
     networkInterfaceName: string,
     tapConfigurationName: string,
     tapConfigurationParameters: NetworkInterfaceTapConfiguration,
     options?: NetworkInterfaceTapConfigurationsCreateOrUpdateOptionalParams
   ): Promise<NetworkInterfaceTapConfigurationsCreateOrUpdateResponse> {
-    const poller = await this.beginCreateOrUpdate(
-      resourceGroupName,
-      networkInterfaceName,
-      tapConfigurationName,
-      tapConfigurationParameters,
-      options
+    return this.client.sendOperationRequest(
+      {
+        resourceGroupName,
+        networkInterfaceName,
+        tapConfigurationName,
+        tapConfigurationParameters,
+        options
+      },
+      createOrUpdateOperationSpec
     );
-    return poller.pollUntilDone();
   }
 
   /**
@@ -401,7 +245,6 @@ const deleteOperationSpec: coreClient.OperationSpec = {
   httpMethod: "DELETE",
   responses: {
     200: {},
-    201: {},
     202: {},
     204: {},
     default: {
@@ -451,12 +294,6 @@ const createOrUpdateOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.NetworkInterfaceTapConfiguration
     },
     201: {
-      bodyMapper: Mappers.NetworkInterfaceTapConfiguration
-    },
-    202: {
-      bodyMapper: Mappers.NetworkInterfaceTapConfiguration
-    },
-    204: {
       bodyMapper: Mappers.NetworkInterfaceTapConfiguration
     },
     default: {