+++ sdk/network/arm-network/src/operationsInterfaces/networkInterfaces.ts
@@ -7,7 +7,6 @@
  */
 
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
-import { SimplePollerLike, OperationState } from "@azure/core-lro";
 import {
   NetworkInterface,
   NetworkInterfacesListCloudServiceRoleInstanceNetworkInterfacesOptionalParams,
@@ -141,18 +140,7 @@ export interface NetworkInterfaces {
    * @param networkInterfaceName The name of the network interface.
    * @param options The options parameters.
    */
-  beginDelete(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    options?: NetworkInterfacesDeleteOptionalParams
-  ): Promise<SimplePollerLike<OperationState<void>, void>>;
-  /**
-   * Deletes the specified network interface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param options The options parameters.
-   */
-  beginDeleteAndWait(
+  delete(
     resourceGroupName: string,
     networkInterfaceName: string,
     options?: NetworkInterfacesDeleteOptionalParams
@@ -175,25 +163,7 @@ export interface NetworkInterfaces {
    * @param parameters Parameters supplied to the create or update network interface operation.
    * @param options The options parameters.
    */
-  beginCreateOrUpdate(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    parameters: NetworkInterface,
-    options?: NetworkInterfacesCreateOrUpdateOptionalParams
-  ): Promise<
-    SimplePollerLike<
-      OperationState<NetworkInterfacesCreateOrUpdateResponse>,
-      NetworkInterfacesCreateOrUpdateResponse
-    >
-  >;
-  /**
-   * Creates or updates a network interface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param parameters Parameters supplied to the create or update network interface operation.
-   * @param options The options parameters.
-   */
-  beginCreateOrUpdateAndWait(
+  createOrUpdate(
     resourceGroupName: string,
     networkInterfaceName: string,
     parameters: NetworkInterface,
@@ -218,23 +188,7 @@ export interface NetworkInterfaces {
    * @param networkInterfaceName The name of the network interface.
    * @param options The options parameters.
    */
-  beginGetEffectiveRouteTable(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    options?: NetworkInterfacesGetEffectiveRouteTableOptionalParams
-  ): Promise<
-    SimplePollerLike<
-      OperationState<NetworkInterfacesGetEffectiveRouteTableResponse>,
-      NetworkInterfacesGetEffectiveRouteTableResponse
-    >
-  >;
-  /**
-   * Gets all route tables applied to a network interface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param options The options parameters.
-   */
-  beginGetEffectiveRouteTableAndWait(
+  getEffectiveRouteTable(
     resourceGroupName: string,
     networkInterfaceName: string,
     options?: NetworkInterfacesGetEffectiveRouteTableOptionalParams
@@ -245,25 +199,7 @@ export interface NetworkInterfaces {
    * @param networkInterfaceName The name of the network interface.
    * @param options The options parameters.
    */
-  beginListEffectiveNetworkSecurityGroups(
-    resourceGroupName: string,
-    networkInterfaceName: string,
-    options?: NetworkInterfacesListEffectiveNetworkSecurityGroupsOptionalParams
-  ): Promise<
-    SimplePollerLike<
-      OperationState<
-        NetworkInterfacesListEffectiveNetworkSecurityGroupsResponse
-      >,
-      NetworkInterfacesListEffectiveNetworkSecurityGroupsResponse
-    >
-  >;
-  /**
-   * Gets all network security groups applied to a network interface.
-   * @param resourceGroupName The name of the resource group.
-   * @param networkInterfaceName The name of the network interface.
-   * @param options The options parameters.
-   */
-  beginListEffectiveNetworkSecurityGroupsAndWait(
+  listEffectiveNetworkSecurityGroups(
     resourceGroupName: string,
     networkInterfaceName: string,
     options?: NetworkInterfacesListEffectiveNetworkSecurityGroupsOptionalParams